Fixed memory access errors.
git-svn-id: http://dminfo:8686/svn/repository/models/branches/dsm2_v8_0@1740 cdc4813b-4270-ec4f-936e-925f93a782c2

+++ src/oprule_interface/dsm2_expressions.cpp
@@ -17,9 +17,9 @@ int register_express_for_data_source(
    data_expr_iter it=std::find(expr_used_as_data.begin(),
                           expr_used_as_data.end(),expr);
    if ( it != expr_used_as_data.end()){
-      return it - expr_used_as_data.begin(); //@todo: is this safe?
+      return (int)(it - expr_used_as_data.begin()); //@todo: is this safe?
    }else{
       expr_used_as_data.push_back(expr);
-      return expr_used_as_data.size()-1;     //@todo: is this safe?
+      return ((int)expr_used_as_data.size())-1;     //@todo: is this safe?
    }
 }
